+++ .changeset/mighty-steaks-tease.md
@@ -2,5 +2,6 @@
 "react-impulse-form": minor
 ---
 
+- Introduce new `isImpulseFormUnit` high order function to check if a value is an instance of `ImpulseFormUnit`.
 - Introduce new `isImpulseFormList` high order function to check if a value is an instance of `ImpulseFormList`.
 - Introduce new `isImpulseFormShape` high order function to check if a value is an instance of `ImpulseFormShape`.
