--- .changeset/tired-snails-crash.md
@@ -1,7 +1,0 @@
----
-"react-impulse-form": minor
----
-
-**BREAKING CHANGES**
-
-Rename `ImpulseFormValue` to `ImpulseFormUnit`.
