+++ packages/react-impulse-form/src/impulse-form-unit/impulse-form-unit-error-setter.ts
@@ -1,0 +1,3 @@
+import type { Setter } from "~/tools/setter"
+
+export type ImpulseFormUnitErrorSetter<TError> = Setter<null | TError>
