+++ packages/react-impulse-form/src/impulse-form-unit/impulse-form-unit-input-setter.ts
@@ -1,0 +1,6 @@
+import type { Setter } from "~/tools/setter"
+
+export type ImpulseFormUnitInputSetter<TInput> = Setter<
+  TInput,
+  [TInput, TInput]
+>
