+++ packages/react-impulse-form/src/impulse-form-unit/impulse-form-unit-validate-on-setter.ts
@@ -1,0 +1,5 @@
+import type { Setter } from "~/tools/setter"
+
+import type { ValidateStrategy } from "../validate-strategy"
+
+export type ImpulseFormUnitValidateOnSetter = Setter<ValidateStrategy>
