+++ packages/react-impulse-form/src/impulse-form-unit/impulse-form-unit-validator.ts
@@ -1,0 +1,5 @@
+import type { Result } from "../result"
+
+export type ImpulseFormUnitValidator<TInput, TError, TOutput> = (
+  input: TInput,
+) => Result<TError, TOutput>
