+++ packages/react-impulse-form/src/impulse-form-unit/is-impulse-form-unit.ts
@@ -1,0 +1,7 @@
+import { ImpulseFormUnit } from "./_impulse-form-unit"
+
+export function isImpulseFormUnit(
+  value: unknown,
+): value is ImpulseFormUnit<unknown> {
+  return value instanceof ImpulseFormUnit
+}
